# book-network-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi Specification-Fawez Teka",
    "description": "Open Api documentation for spring security",
    "termsOfService": "Terms Of services",
    "contact": { "name": "Fawez", "email": "tekafawez@gmail.com" },
    "license": { "name": "license name", "url": "https://some-url.com" },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8088/api/v1", "description": "Local ENV" },
    { "url": "http://fawezteka.com/dev", "description": "Prod ENV" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/bookTransactionHistories/{id}/book/{propertyId}": {
      "get": {
        "tags": ["book-transaction-history-property-reference-controller"],
        "description": "get-book-by-booktransactionhistory-Id",
        "operationId": "followPropertyReference-booktransactionhistory-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "tags": ["book-transaction-history-property-reference-controller"],
        "description": "delete-book-by-booktransactionhistory-Id",
        "operationId": "deletePropertyReferenceId-booktransactionhistory-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    },
...
    "/books": {
      "get": {
        "tags": ["book-entity-controller"],
        "description": "get-book",
        "operationId": "getCollectionResource-book-get_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": { "minimum": 0, "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": { "minimum": 1, "type": "integer", "default": 20 }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelBook"
                }
              },
              "application/x-spring-data-compact+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelBook"
                }
              },
              "text/uri-list": { "schema": { "type": "string" } }
            }
          }
        }
      },
      "post": {
        "tags": ["book", "book-entity-controller"],
        "summary": "Save a book",
        "description": "create-book",
        "operationId": "postCollectionResource-book-post_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          }
        }
      }
    },
...
    "/books/{id}": {
      "get": {
        "tags": ["book-entity-controller"],
        "description": "get-book",
        "operationId": "getItemResource-book-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "tags": ["book-entity-controller"],
        "description": "update-book",
        "operationId": "putItemResource-book-put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookRequestBody" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "204": { "description": "No Content" }
        }
      },
      "delete": {
        "tags": ["book-entity-controller"],
        "description": "delete-book",
        "operationId": "deleteItemResource-book-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "tags": ["book-entity-controller"],
        "description": "patch-book",
        "operationId": "patchItemResource-book-patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookRequestBody" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "204": { "description": "No Content" }
        }
      }
    },
...
    "/books/{id}/owner": {
      "get": {
        "tags": ["book-property-reference-controller"],
        "description": "get-user-by-book-Id",
        "operationId": "followPropertyReference-book-get_3_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelUser" }
              },
              "text/uri-list": { "schema": { "type": "string" } }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "tags": ["book-property-reference-controller"],
        "description": "update-user-by-book-Id",
        "operationId": "createPropertyReference-book-put_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CollectionModelObject" }
            },
            "application/x-spring-data-compact+json": {
              "schema": { "$ref": "#/components/schemas/CollectionModelObject" }
            },
            "text/uri-list": { "schema": { "type": "string" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelUser" }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelUser" }
              }
            }
          },
          "204": { "description": "No Content" }
        }
      },
      "delete": {
        "tags": ["book-property-reference-controller"],
        "description": "delete-user-by-book-Id",
        "operationId": "deletePropertyReference-book-delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "tags": ["book-property-reference-controller"],
        "description": "patch-user-by-book-Id",
        "operationId": "createPropertyReference-book-patch_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CollectionModelObject" }
            },
            "application/x-spring-data-compact+json": {
              "schema": { "$ref": "#/components/schemas/CollectionModelObject" }
            },
            "text/uri-list": { "schema": { "type": "string" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelUser" }
              }
            }
          },
          "204": { "description": "No Content" }
        }
      }
    },
...
    "/feedbacks/{id}/book/{propertyId}": {
      "get": {
        "tags": ["feedback-property-reference-controller"],
        "description": "get-book-by-feedback-Id",
        "operationId": "followPropertyReference-feedback-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "tags": ["feedback-property-reference-controller"],
        "description": "delete-book-by-feedback-Id",
        "operationId": "deletePropertyReferenceId-feedback-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    },
...
    "/users/{id}/books/{propertyId}": {
      "get": {
        "tags": ["user-property-reference-controller"],
        "description": "get-book-by-user-Id",
        "operationId": "followPropertyReference-user-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/CollectionModelBook" }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "tags": ["user-property-reference-controller"],
        "description": "delete-book-by-user-Id",
        "operationId": "deletePropertyReferenceId-user-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    },
...
    "/books/cover/{book-id}": {
      "post": {
        "tags": ["book"],
        "operationId": "uploadBookCoverPicture",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
    "/books/borrow/{book-id}": {
      "post": {
        "tags": ["book"],
        "operationId": "borrowBook",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
...
    "/books/shareable/{book-id}": {
      "patch": {
        "tags": ["book"],
        "operationId": "updateShareableStatus",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/books/borrow/return/{book-id}": {
      "patch": {
        "tags": ["book"],
        "operationId": "returnBorrowBook",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/books/borrow/return/approve/{book-id}": {
      "patch": {
        "tags": ["book"],
        "operationId": "approveReturnBorrowBook",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/books/archived/{book-id}": {
      "patch": {
        "tags": ["book"],
        "operationId": "updateArchivedStatus",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "type": "integer", "format": "int32" }
              }
            }
          }
        }
      }
    },
    "/feedback/book/{book-id}": {
      "get": {
        "tags": ["Feedback"],
        "operationId": "findAllFeedbacksByBook",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book-id}": {
      "get": {
        "tags": ["book"],
        "operationId": "findBookById",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": { "$ref": "#/components/schemas/BookResponse" }
              }
            }
          }
        }
      }
    },
    "/books/returned": {
      "get": {
        "tags": ["book"],
        "operationId": "findAllReturnedBooks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/owner": {
      "get": {
        "tags": ["book"],
        "operationId": "findAllBooksByOwner",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/findAllBooks": {
      "get": {
        "tags": ["book"],
        "operationId": "findAllBooks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/borrowed": {
      "get": {
        "tags": ["book"],
        "operationId": "findAllBorrowedBooks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": { "$ref": "#/components/schemas/Link" }
      },
...
      "EntityModelUser": {
        "type": "object",
        "properties": {
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "dateOfBirth": { "type": "string", "format": "date" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "accountLocked": { "type": "boolean" },
          "enabled": { "type": "boolean" },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" },
          "name": { "type": "string" },
          "accountNonLocked": { "type": "boolean" },
          "accountNonExpired": { "type": "boolean" },
          "credentialsNonExpired": { "type": "boolean" },
          "username": { "type": "string" },
          "authorities": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GrantedAuthority" }
          },
          "_links": { "$ref": "#/components/schemas/Links" }
        }
      },
      "CollectionModelObject": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "objects": { "type": "array", "items": { "type": "object" } }
            }
          },
          "_links": { "$ref": "#/components/schemas/Links" }
        }
      },
      "EntityModelBook": {
        "type": "object",
        "properties": {
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" },
          "createdBy": { "type": "integer", "format": "int32" },
          "lastModifiedBy": { "type": "integer", "format": "int32" },
          "title": { "type": "string" },
          "authorName": { "type": "string" },
          "isbn": { "type": "string" },
          "synopsis": { "type": "string" },
          "bookCover": { "type": "string" },
          "archived": { "type": "boolean" },
          "shareable": { "type": "boolean" },
          "rate": { "type": "number", "format": "double" },
          "_links": { "$ref": "#/components/schemas/Links" }
        }
      },
      "PagedModelEntityModelBook": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "books": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/EntityModelBook" }
              }
            }
          },
          "_links": { "$ref": "#/components/schemas/Links" },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
...
      "CollectionModelBook": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "books": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/BookResponse" }
              }
            }
          },
          "_links": { "$ref": "#/components/schemas/Links" }
        }
      },
...
      "BookRequestBody": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" },
          "createdBy": { "type": "integer", "format": "int32" },
          "lastModifiedBy": { "type": "integer", "format": "int32" },
          "title": { "type": "string" },
          "authorName": { "type": "string" },
          "isbn": { "type": "string" },
          "synopsis": { "type": "string" },
          "bookCover": { "type": "string" },
          "archived": { "type": "boolean" },
          "shareable": { "type": "boolean" },
          "owner": { "type": "string" },
          "feedbacks": { "type": "array", "items": { "type": "string" } },
          "histories": { "type": "array", "items": { "type": "string" } },
          "rate": { "type": "number", "format": "double" }
        }
      },
...
      "BookResponse": {
        "type": "object",
        "properties": {
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" },
          "createdBy": { "type": "integer", "format": "int32" },
          "lastModifiedBy": { "type": "integer", "format": "int32" },
          "title": { "type": "string" },
          "authorName": { "type": "string" },
          "isbn": { "type": "string" },
          "synopsis": { "type": "string" },
          "bookCover": { "type": "string" },
          "archived": { "type": "boolean" },
          "shareable": { "type": "boolean" },
          "rate": { "type": "number", "format": "double" }
        }
      },
...
      "BookRequest": {
        "required": ["authorName", "isbn", "synopsis", "title"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int32" },
          "title": { "type": "string" },
          "authorName": { "type": "string" },
          "isbn": { "type": "string" },
          "synopsis": { "type": "string" },
          "shareable": { "type": "boolean" }
        }
      },
...
      "PageResponseFeedbackResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FeedbackResponse" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
...
      "PageResponseBorrowedBookResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BorrowedBookResponse" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
      "PageResponseBookResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BookResponse" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Jwt auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}